feat: added optional filters on mock resource search (#16)
This commit contains a new functionality made on mock resource research operation. This new functionality aims to help the pagoPA Shared Toolbox portal on search operation, in order to guarantee a better user experience.  
Also, the automatic update of Terraform resources with dedicated GitHub pipeline is removed because there is an issue with the use of wrong authorization token, caused by the migration of infrastructure's apply strategy to managed-identity mode.

#### List of Changes
 - Added optional filters for filtering elements during mock resource read-all invocation
 - Temporarily removed Terraform's OpenAPI automatic update